pyproject.toml
[project]
name = "document-qa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Terry Zhao", email = "terry@runllama.ai" }
]
requires-python = ">=3.12"
dependencies = [
    "llama-index-workflows>=2.2.0",
    "python-cowsay>=1.2.1",
    "llama-cloud-services>=0.6.0",
    "llama-index-core>=0.12.0",
    "llama-index-llms-openai>=0.3.0",
    "llama-index-embeddings-openai>=0.3.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []
